test: 임의의 제보 등록에 대한 테스트 코드 추가 및 위경도에 대한 float값을 double로 변경

+++ src/main/java/com/ktc/togetherPet/model/dto/report/ReportCreateRequestDTO.java
@@ -11,9 +11,9 @@ public record ReportCreateRequestDTO(
     @NotNull
     String color,
     @NotNull
-    Float foundLatitude,
+    double foundLatitude,
     @NotNull
-    Float foundLongitude,
+    double foundLongitude,
     @NotNull
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
     LocalDateTime foundDate,
